+++ cpp/benchmarks/fixture/templated_benchmark_fixture.hpp
