+++ cpp/benchmarks/hashing/hash.cpp
@@ -15,7 +15,7 @@
  */
 
 #include <benchmark/benchmark.h>
-#include <benchmarks/common/generate_benchmark_input.hpp>
+#include <benchmarks/common/generate_input.hpp>
 #include <benchmarks/fixture/benchmark_fixture.hpp>
 #include <benchmarks/synchronization/synchronization.hpp>
