+++ cpp/benchmarks/io/orc/orc_writer.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2021, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2022, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,26 +14,27 @@
  * limitations under the License.
  */
 
+#include "cudf/io/types.hpp"
 #include <benchmark/benchmark.h>
 
-#include <benchmarks/common/generate_benchmark_input.hpp>
+#include <benchmarks/common/generate_input.hpp>
 #include <benchmarks/fixture/benchmark_fixture.hpp>
-#include <benchmarks/io/cuio_benchmark_common.hpp>
+#include <benchmarks/io/cuio_common.hpp>
 #include <benchmarks/synchronization/synchronization.hpp>
 
-#include <cudf/io/parquet.hpp>
+#include <cudf/io/orc.hpp>
 
 // to enable, run cmake with -DBUILD_BENCHMARKS=ON
 
-constexpr size_t data_size         = 512 << 20;
+constexpr int64_t data_size        = 512 << 20;
 constexpr cudf::size_type num_cols = 64;
 
 namespace cudf_io = cudf::io;
 
-class ParquetWrite : public cudf::benchmark {
+class OrcWrite : public cudf::benchmark {
 };
 
-void BM_parq_write_varying_inout(benchmark::State& state)
+void BM_orc_write_varying_inout(benchmark::State& state)
 {
   auto const data_types             = get_type_or_group(state.range(0));
   cudf::size_type const cardinality = state.range(1);
@@ -53,21 +54,26 @@ void BM_parq_write_varying_inout(benchmark::State& state)
   auto mem_stats_logger = cudf::memory_stats_logger();
   for (auto _ : state) {
     cuda_event_timer raii(state, true);  // flush_l2_cache = true, stream = 0
-    cudf_io::parquet_writer_options opts =
-      cudf_io::parquet_writer_options::builder(source_sink.make_sink_info(), view)
+    cudf_io::orc_writer_options options =
+      cudf_io::orc_writer_options::builder(source_sink.make_sink_info(), view)
         .compression(compression);
-    cudf_io::write_parquet(opts);
+    cudf_io::write_orc(options);
   }
 
   state.SetBytesProcessed(data_size * state.iterations());
   state.counters["peak_memory_usage"] = mem_stats_logger.peak_memory_usage();
 }
 
-void BM_parq_write_varying_options(benchmark::State& state)
+void BM_orc_write_varying_options(benchmark::State& state)
 {
-  auto const compression  = static_cast<cudf::io::compression_type>(state.range(0));
-  auto const enable_stats = static_cast<cudf::io::statistics_freq>(state.range(1));
-  auto const file_path    = state.range(2) != 0 ? "unused_path.parquet" : "";
+  auto const compression = static_cast<cudf::io::compression_type>(state.range(0));
+  auto const stats_freq  = [&] {
+    switch (state.range(2)) {
+      case 0: return cudf::io::STATISTICS_NONE;
+      case 1: return cudf::io::ORC_STATISTICS_STRIPE;
+      default: return cudf::io::ORC_STATISTICS_ROW_GROUP;
+    }
+  }();
 
   auto const data_types = get_type_or_group({int32_t(type_group_id::INTEGRAL_SIGNED),
                                              int32_t(type_group_id::FLOATING_POINT),
@@ -83,41 +89,39 @@ void BM_parq_write_varying_options(benchmark::State& state)
   auto mem_stats_logger = cudf::memory_stats_logger();
   for (auto _ : state) {
     cuda_event_timer raii(state, true);  // flush_l2_cache = true, stream = 0
-    cudf_io::parquet_writer_options const options =
-      cudf_io::parquet_writer_options::builder(source_sink.make_sink_info(), view)
+    cudf_io::orc_writer_options const options =
+      cudf_io::orc_writer_options::builder(source_sink.make_sink_info(), view)
         .compression(compression)
-        .stats_level(enable_stats)
-        .column_chunks_file_paths({file_path});
-    cudf_io::write_parquet(options);
+        .enable_statistics(stats_freq);
+    cudf_io::write_orc(options);
   }
 
   state.SetBytesProcessed(data_size * state.iterations());
   state.counters["peak_memory_usage"] = mem_stats_logger.peak_memory_usage();
 }
 
-#define PARQ_WR_BM_INOUTS_DEFINE(name, type_or_group, sink_type)                              \
-  BENCHMARK_DEFINE_F(ParquetWrite, name)                                                      \
-  (::benchmark::State & state) { BM_parq_write_varying_inout(state); }                        \
-  BENCHMARK_REGISTER_F(ParquetWrite, name)                                                    \
+#define ORC_WR_BM_INOUTS_DEFINE(name, type_or_group, sink_type)                               \
+  BENCHMARK_DEFINE_F(OrcWrite, name)                                                          \
+  (::benchmark::State & state) { BM_orc_write_varying_inout(state); }                         \
+  BENCHMARK_REGISTER_F(OrcWrite, name)                                                        \
     ->ArgsProduct({{int32_t(type_or_group)}, {0, 1000}, {1, 32}, {true, false}, {sink_type}}) \
     ->Unit(benchmark::kMillisecond)                                                           \
     ->UseManualTime();
 
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, integral, type_group_id::INTEGRAL);
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, floats, type_group_id::FLOATING_POINT);
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, decimal, type_group_id::FIXED_POINT);
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, timestamps, type_group_id::TIMESTAMP);
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, string, cudf::type_id::STRING);
-WR_BENCHMARK_DEFINE_ALL_SINKS(PARQ_WR_BM_INOUTS_DEFINE, list, cudf::type_id::LIST);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, integral, type_group_id::INTEGRAL_SIGNED);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, floats, type_group_id::FLOATING_POINT);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, decimal, type_group_id::FIXED_POINT);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, timestamps, type_group_id::TIMESTAMP);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, string, cudf::type_id::STRING);
+WR_BENCHMARK_DEFINE_ALL_SINKS(ORC_WR_BM_INOUTS_DEFINE, list, cudf::type_id::LIST);
 
-BENCHMARK_DEFINE_F(ParquetWrite, writer_options)
-(::benchmark::State& state) { BM_parq_write_varying_options(state); }
-BENCHMARK_REGISTER_F(ParquetWrite, writer_options)
+BENCHMARK_DEFINE_F(OrcWrite, writer_options)
+(::benchmark::State& state) { BM_orc_write_varying_options(state); }
+BENCHMARK_REGISTER_F(OrcWrite, writer_options)
   ->ArgsProduct({{int32_t(cudf::io::compression_type::NONE),
                   int32_t(cudf::io::compression_type::SNAPPY)},
-                 {int32_t(cudf::io::statistics_freq::STATISTICS_NONE),
-                  int32_t(cudf::io::statistics_freq::STATISTICS_ROWGROUP),
-                  int32_t(cudf::io::statistics_freq::STATISTICS_PAGE)},
-                 {false, true}})
+                 {int32_t{cudf::io::STATISTICS_NONE},
+                  int32_t{cudf::io::ORC_STATISTICS_STRIPE},
+                  int32_t{cudf::io::ORC_STATISTICS_ROW_GROUP}}})
   ->Unit(benchmark::kMillisecond)
   ->UseManualTime();
