+++ cpp/benchmarks/sort/sort.cpp
@@ -23,7 +23,7 @@
 #include <cudf_test/table_utilities.hpp>
 
 #include <benchmark/benchmark.h>
-#include <benchmarks/common/generate_benchmark_input.hpp>
+#include <benchmarks/common/generate_input.hpp>
 #include <benchmarks/fixture/benchmark_fixture.hpp>
 #include <benchmarks/synchronization/synchronization.hpp>
