+++ cpp/benchmarks/string/convert_fixed_point.cpp
@@ -18,7 +18,7 @@
 #include <synchronization/synchronization.hpp>
 
 #include <benchmark/benchmark.h>
-#include <benchmarks/common/generate_benchmark_input.hpp>
+#include <benchmarks/common/generate_input.hpp>
 
 #include <cudf/strings/convert/convert_fixed_point.hpp>
 #include <cudf/strings/convert/convert_floats.hpp>
