+++ cpp/benchmarks/string/translate.cpp
@@ -17,7 +17,7 @@
 #include "string_bench_args.hpp"
 
 #include <benchmark/benchmark.h>
-#include <benchmarks/common/generate_benchmark_input.hpp>
+#include <benchmarks/common/generate_input.hpp>
 #include <benchmarks/fixture/benchmark_fixture.hpp>
 #include <benchmarks/synchronization/synchronization.hpp>
